refactor: remove `_Workflow`
The implementation of `Experiment` using the `_Workflow` class to manage the data and methods order is over-complicated. It prevents the new feature that users could pass in a meta-property table instead of glycan structures. This commit refactored the `Experiment` class, to use the simple way of mananging data and method orders.